requirements.txt
openai==1.3.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
rembg==2.0.61
pytest==7.4.3
httpx==0.25.2
onnxruntime==1.20.1
